refactor: move HistoryTrimHook into its own example
HistoryTrimHook / Session.get_framework_session() doesn't belong bundled into
the general hooks demo (GuardRail/RAG/Disclaimer) - it's a separate concept
worth its own example. Move it, and its unit test, into a new
examples/api/session-context/, and update doc references
(session.md, hooks.md, ak-add-capabilities/SKILL.md) and the e2e CI test
config accordingly. No behavioral changes in this commit - HistoryTrimHook's
logic and its test are carried over unchanged.

Addresses PR #622 review comment: https://github.com/yaalalabs/agent-kernel/pull/622#discussion_r3776139405
