+++ .github/test-config.yaml
@@ -80,6 +80,8 @@ e2e:
     - type: api
       path: examples/api/hooks
     - type: api
+      path: examples/api/session-context
+    - type: api
       path: examples/api/slack
       requires_slack: true
     - type: api
