+++ ak-py/src/agentkernel/skills/ak-add-capabilities/SKILL.md
@@ -602,8 +602,8 @@ class HistoryTrimHook(PostHook):
         return "HistoryTrimHook"
 ```
 
-See `examples/api/hooks/hooks.py` (`HistoryTrimHook`) for a complete example that caps the OpenAI Agents
-SDK's raw conversation history after every turn.
+See `examples/api/session-context/hooks.py` (`HistoryTrimHook`) for a complete example that caps the
+OpenAI Agents SDK's raw conversation history after every turn.
 
 ---
