+++ examples/api/session-context/app.py
@@ -2,7 +2,7 @@ from agentkernel.api import RESTAPI
 from agentkernel.openai import OpenAIModule
 from agents import Agent
 
-from hooks import DisclaimerHook, GuardRailHook, RAGHook
+from hooks import HistoryTrimHook
 
 # Create a simple question-answering agent
 qa_agent = Agent(
@@ -14,10 +14,9 @@ qa_agent = Agent(
 )
 
 # Register the agent with the OpenAI module
-# Register pre-execution hooks in order: RAG hook first (to inject context), then GuardRail (to validate input)
-# Register post-execution hooks to add disclaimer to responses
-# The hooks will be executed in the order they are provided
-OpenAIModule([qa_agent]).pre_hook(qa_agent, [RAGHook(), GuardRailHook()]).post_hook(qa_agent, [DisclaimerHook()])
+# Register a post-execution hook that caps the framework-native session history so it
+# never grows unbounded
+OpenAIModule([qa_agent]).post_hook(qa_agent, [HistoryTrimHook()])
 
 if __name__ == "__main__":
     RESTAPI.run()
