+++ examples/api/session-context/build.sh
@@ -1,0 +1,16 @@
+#!/bin/bash
+
+set -euo pipefail
+if command -v pyenv >/dev/null 2>&1; then
+  uv venv --python "$(pyenv which python)" --allow-existing
+else
+  uv venv --allow-existing
+fi
+
+if [[ ${1-} != "local" ]]; then
+  uv sync --all-extras
+else
+  # For local development of agentkernel, you can force reinstall from local dist
+  uv sync --find-links ../../../ak-py/dist --all-extras
+  uv pip install --force-reinstall --no-deps --no-index --find-links ../../../ak-py/dist agentkernel[api,openai,test] || true
+fi
