+++ examples/api/session-context/hooks_test.py
