+++ examples/api/session-context/pyproject.toml
@@ -1,0 +1,35 @@
+[project]
+name = "openai-session-context-demo"
+version = "0.1.0"
+description = "Demo of Session.get_framework_session() via a history-trimming post-hook with OpenAI Agents in Agent Kernel"
+readme = "README.md"
+requires-python = ">=3.12"
+dependencies = [
+    "agentkernel[openai,api]>=0.8.1",
+]
+
+[dependency-groups]
+dev = [
+    "agentkernel[test]>=0.8.1",
+    "black==26.5.1",
+    "isort>=5.0.0",
+    "mypy>=1.0.0",
+]
+
+[tool.uv]
+package = false
+
+[tool.isort]
+profile = "black"
+line_length = 120
+
+[tool.mypy]
+python_version = "3.12"
+warn_return_any = true
+warn_unused_configs = true
+disallow_untyped_defs = false
+disallow_incomplete_defs = false
+
+[tool.black]
+line-length = 120
+target-version = ["py312"]
